+performStaircaseGapCheck(player)
If no ground is found directly beneath a killed player they might be
standing over a gap between 2 staircase steps.

So this new function will try to find one of these steps and if it does
then an RE will spawn at the player's death location instead of at a
player spawn point.